Fix Drafts mode ghost for __MCP_CommandListener during playtests
Install the nav bridge script on the play-server peer only so start_playtest no longer creates ephemeral scripts in edit mode that Drafts mode keeps after Destroy().
